feat: allow start again edition
