+++ components/Editor.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { computed, onMounted, ref, type Component } from "vue";
+import { computed, type Component } from "vue";
 
 import { storeToRefs } from "pinia";
 import {
@@ -19,16 +19,13 @@ import {
   UserIcon,
   WrenchIcon,
 } from "@heroicons/vue/16/solid";
-import type { Asset, Category, Experience, Export } from "@/types";
+import type { Category } from "@/types";
 import { useEditorStore } from "@/stores/editor";
-import { useLetterStore } from "@/stores/letter";
 import { useProfileStore } from "@/stores/profile";
 import { useResumeStore } from "@/stores/resume";
 import { discouragedLayoutTemplates, fixedLayoutTemplates } from "@/globals";
 import { getEntryHeading } from "@/utils/editor";
-import { getRandomAsset, getRandomExperience } from "@/utils/random";
 import { capitalize } from "@/utils/string";
-import useDialog from "~/composables/use-dialog";
 import EmailEditor from "@/fragments/EmailEditor.vue";
 import LetterEditor from "@/fragments/LetterEditor.vue";
 import ProfileEditor from "@/fragments/ProfileEditor.vue";
@@ -44,17 +41,6 @@ const { documentType } = storeToRefs(useEditorStore());
 const { template } = storeToRefs(useProfileStore());
 
 const { categories } = storeToRefs(useResumeStore());
-
-const profileStore = useProfileStore();
-const profile = storeToRefs(profileStore);
-const resumeStore = useResumeStore();
-const resume = storeToRefs(resumeStore);
-const letterStore = useLetterStore();
-const letter = storeToRefs(letterStore);
-
-const { dialog, openModal, closeModal } = useDialog();
-
-const isImportError = ref(false);
 
 const isLayoutDisabled = computed(() =>
   fixedLayoutTemplates.includes(template.value),
@@ -73,128 +59,6 @@ const discouragedLayoutText = computed(() => {
   return `${capitalize(layouts)} ${isPlural ? "layouts are" : "layout is"} discouraged for this template.`;
 });
 
-function generateStores() {
-  // Profile
-  profile.name.value = "Firstname Lastname";
-  profile.title.value = "Title";
-  profile.about.value = "About";
-  profile.contactDetails.value = [
-    { type: "personal", icon: "drivingLicense", value: "Driving license" },
-    {
-      type: "personal",
-      icon: "address",
-      value: "Streetnumber street Streetname, ZIPCODE City Country",
-    },
-    { type: "personal", icon: "email", value: "email@email.com" },
-    { type: "personal", icon: "phone", value: "061122334455" },
-    { type: "social", icon: "gitHub", value: "github.com" },
-    { type: "social", icon: "linkedIn", value: "linkedin.com" },
-    { type: "social", icon: "viadeo", value: "viadeo.com" },
-  ];
-
-  // Resume
-  resume.categories.value = [
-    {
-      nature: "experience",
-      type: "work",
-      name: "Work experience",
-      entries: [
-        getRandomExperience("work" as Experience["type"]),
-        getRandomExperience("work" as Experience["type"]),
-        getRandomExperience("work" as Experience["type"]),
-      ],
-      layout: "full",
-      isVisible: true,
-    },
-    {
-      nature: "experience",
-      type: "education",
-      name: "Education",
-      entries: [
-        getRandomExperience("education" as Experience["type"]),
-        getRandomExperience("education" as Experience["type"]),
-      ],
-      layout: "full",
-      isVisible: true,
-    },
-    {
-      nature: "experience",
-      type: "certificate",
-      name: "Certificates",
-      entries: [getRandomExperience("certificate" as Experience["type"])],
-      layout: "half",
-      isVisible: true,
-    },
-    {
-      nature: "experience",
-      type: "publication",
-      name: "Publications",
-      entries: [
-        getRandomExperience("publication" as Experience["type"]),
-        getRandomExperience("publication" as Experience["type"]),
-      ],
-      layout: "half",
-      isVisible: true,
-    },
-    {
-      nature: "experience",
-      type: "award",
-      name: "Awards",
-      entries: [getRandomExperience("award" as Experience["type"])],
-      layout: "half",
-      isVisible: true,
-    },
-    {
-      nature: "asset",
-      type: "skill",
-      name: "Skills",
-      entries: [
-        getRandomAsset("skill" as Asset["type"]),
-        getRandomAsset("skill" as Asset["type"]),
-        getRandomAsset("skill" as Asset["type"]),
-      ],
-      layout: "half",
-      isVisible: true,
-    },
-    {
-      nature: "asset",
-      type: "language",
-      name: "Languages",
-      entries: [
-        getRandomAsset("language" as Asset["type"]),
-        getRandomAsset("language" as Asset["type"]),
-        getRandomAsset("language" as Asset["type"]),
-      ],
-      layout: "half",
-      isVisible: true,
-    },
-  ];
-
-  // Letter
-  letter.senderDetails.value = [
-    "Firstname Lastname",
-    "Streetnumber street Streetname",
-    "ZIPCODE City Country",
-  ];
-  letter.recipientDetails.value = [
-    "Firstname Lastname",
-    "Streetnumber street Streetname",
-    "ZIPCODE City Country",
-  ];
-  letter.subject.value = "Cover letter subject";
-  letter.reference.value = "Ref.: Advertisement reference";
-  letter.paragraphs.value = [
-    "Salutation,",
-    "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-    "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?",
-    "At vero eos et accusamus et iusto odio dignissimos ducimus qui blanditiis praesentium voluptatum deleniti atque corrupti quos dolores et quas molestias excepturi sint occaecati cupiditate non provident, similique sunt in culpa qui officia deserunt mollitia animi, id est laborum et dolorum fuga. Et harum quidem rerum facilis est et expedita distinctio. Nam libero tempore, cum soluta nobis est eligendi optio cumque nihil impedit quo minus id quod maxime placeat facere possimus, omnis voluptas assumenda est, omnis dolor repellendus. Temporibus autem quibusdam et aut officiis debitis aut rerum necessitatibus saepe eveniet ut et voluptates repudiandae sint et molestiae non recusandae. Itaque earum rerum hic tenetur a sapiente delectus, ut aut reiciendis voluptatibus maiores alias consequatur aut perferendis doloribus asperiores repellat.",
-    "Outroduction.",
-    "Regards.",
-  ];
-
-  closeModal();
-}
-
 function getCategoryIcon(categoryType: Category["type"]) {
   const iconMapper: { [key in Category["type"]]: Component } = {
     award: TrophyIcon,
@@ -210,184 +74,10 @@ function getCategoryIcon(categoryType: Category["type"]) {
   };
   return iconMapper[categoryType];
 }
-
-function importFromJson(event: Event) {
-  isImportError.value = false;
-  try {
-    // @ts-expect-error It seems there is no default <input type=file /> native TS type...
-    const file = event.currentTarget?.files[0];
-
-    const fileReader = new FileReader();
-    fileReader.readAsText(file, "UTF-8");
-    fileReader.onload = function (fileReaderEvent) {
-      if (!fileReaderEvent.target?.result) {
-        isImportError.value = true;
-        return;
-      }
-
-      const toImport: Export = JSON.parse(
-        fileReaderEvent.target.result.toString(),
-      );
-      if (!toImport.isNiceResumeExport) {
-        isImportError.value = true;
-        return;
-      }
-
-      Object.entries(toImport).forEach(([key, value]) => {
-        // @ts-expect-error Build object on the fly
-        if (letter[key]) {
-          // @ts-expect-error Build object on the fly
-          letter[key].value = value;
-        }
-        // @ts-expect-error Build object on the fly
-        if (profile[key]) {
-          // @ts-expect-error Build object on the fly
-          profile[key].value = value;
-        }
-        // @ts-expect-error Build object on the fly
-        if (resume[key]) {
-          // @ts-expect-error Build object on the fly
-          resume[key].value = value;
-        }
-      });
-
-      closeModal();
-    };
-    fileReader.onerror = function () {
-      isImportError.value = true;
-    };
-  } catch {
-    isImportError.value = true;
-  }
-}
-
-function importFromJsonResume(event: Event) {
-  isImportError.value = false;
-  try {
-    // @ts-expect-error It seems there is no default <input type=file /> native TS type...
-    const file = event.currentTarget?.files[0];
-
-    const fileReader = new FileReader();
-    fileReader.readAsText(file, "UTF-8");
-    fileReader.onload = function (fileReaderEvent) {
-      if (!fileReaderEvent.target?.result) {
-        isImportError.value = true;
-        return;
-      }
-
-      const toImport: Export = JSON.parse(
-        fileReaderEvent.target.result.toString(),
-      );
-
-      // TODO start user-journey to apply mapping
-      // TODO improve mapping (startDate/endDate->period)
-      // TODO improve mapping (highlights/tags->handle)
-      // TODO improve mapping (references->lost)
-
-      Object.entries(toImport).forEach(([key, value]) => {
-        // @ts-expect-error Build object on the fly
-        if (letter[key]) {
-          // @ts-expect-error Build object on the fly
-          letter[key].value = value;
-        }
-        // @ts-expect-error Build object on the fly
-        if (profile[key]) {
-          // @ts-expect-error Build object on the fly
-          profile[key].value = value;
-        }
-        // @ts-expect-error Build object on the fly
-        if (resume[key]) {
-          // @ts-expect-error Build object on the fly
-          resume[key].value = value;
-        }
-      });
-
-      closeModal();
-    };
-    fileReader.onerror = function () {
-      isImportError.value = true;
-    };
-  } catch {
-    isImportError.value = true;
-  }
-}
-
-function resetStores() {
-  profileStore.$reset();
-  resumeStore.$reset();
-  letterStore.$reset();
-  closeModal();
-}
-
-onMounted(() => {
-  openModal();
-});
 </script>
 
 <template>
   <div class="print:hidden flex flex-col xl:w-[calc(100%-210mm)] h-screen">
-    <dialog ref="dialog" class="dialog max-w-screen-sm">
-      <p class="mb-8 text-center text-2xl font-bold text-pink-500">
-        How do you want to start editing?
-      </p>
-      <div class="flex flex-col gap-4">
-        <button class="button bgGradient p-[2px]" @click="closeModal">
-          <div class="button bg-white h-full w-full rounded-sm">
-            <span class="textGradient">Continue where I left off</span>
-          </div>
-        </button>
-        <button class="button bgGradient p-[2px]" @click="resetStores">
-          <div class="button bg-white h-full w-full rounded-sm">
-            <span class="textGradient">Start from scratch</span>
-          </div>
-        </button>
-        <button class="button bgGradient p-[2px]" @click="generateStores">
-          <div class="button bg-white h-full w-full rounded-sm">
-            <span class="textGradient">Edit pre-filled data</span>
-          </div>
-        </button>
-        <label
-          for="editorSaveFileReader"
-          class="button bgGradient p-[2px] cursor-pointer"
-        >
-          <div class="button bg-white h-full w-full rounded-sm">
-            <span class="textGradient">
-              Import a save file from a previous session
-            </span>
-            <input
-              id="editorSaveFileReader"
-              class="hidden"
-              type="file"
-              accept=".json"
-              @change="importFromJson"
-            />
-          </div>
-        </label>
-        <label
-          for="editorJsonResumeFileReader"
-          class="button bgGradient p-[2px] cursor-pointer"
-        >
-          <div class="button bg-white h-full w-full rounded-sm">
-            <span class="textGradient">Import a JSON Resume file*</span>
-            <input
-              id="editorJsonResumeFileReader"
-              class="hidden"
-              type="file"
-              accept=".json"
-              @change="importFromJsonResume"
-            />
-          </div>
-        </label>
-        <p class="text-blue-500 text-center">
-          *Full compatibility will be soon available. In The meantime,
-          double-check dates, highlights and tags after import, and be informed
-          that profile image and references are not supported yet.
-        </p>
-        <p v-if="isImportError" class="text-red-500 text-center">
-          Error while importing data from local file.
-        </p>
-      </div>
-    </dialog>
     <TheHeader />
     <div class="flex">
       <aside>
